speedtrade/backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic[email]==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Async & Caching
aioredis==2.0.1
redis==5.0.1
celery==5.3.4

# Trading Libraries (from awesome-quant)
alpaca-py>=0.30.0
ccxt==4.1.54
yfinance==0.2.32
pandas>=2.0.0
numpy>=1.24.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
bcrypt==4.1.1

# Monitoring & Logging
loguru==0.7.2
sentry-sdk==1.38.0
prometheus-client==0.19.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
faker==20.1.0

# Utilities
python-dotenv==1.0.0
websockets==10.4
